clean up helpers function files
