+++ helpers/displayRoom.js
